Reorganize utils into their own file, add specs
