+++ context/app/static/js/components/detailPage/provenance/hooks.ts
@@ -2,106 +2,10 @@ import { getAuthHeader } from 'js/helpers/functions';
 import { useAppContext } from 'js/components/Contexts';
 import useSWR from 'swr';
 import { ProvData } from 'js/components/detailPage/provenance/types';
-import { SearchRequest } from '@elastic/elasticsearch/lib/api/types';
 import { fetchSearchData } from 'js/hooks/useSearchData';
+import { getAncestorsQuery, getDescendantsQuery } from './queries';
+import { createProvDataURL, nonDestructiveMerge } from './utils';
 
-const getAncestorsQuery = (uuid: string) => ({
-  query: {
-    bool: {
-      must: [
-        {
-          term: {
-            'entity_type.keyword': 'Dataset',
-          },
-        },
-        {
-          term: {
-            'processing.keyword': 'raw',
-          },
-        },
-        {
-          term: {
-            descendant_ids: uuid,
-          },
-        },
-        {
-          terms: {
-            'mapped_status.keyword': ['QA', 'Published'],
-          },
-        },
-        {
-          bool: {
-            must_not: {
-              exists: {
-                field: 'next_revision_uuid',
-              },
-            },
-          },
-        },
-        {
-          bool: {
-            must_not: {
-              exists: {
-                field: 'sub_status',
-              },
-            },
-          },
-        },
-        {
-          bool: {
-            must_not: {
-              term: {
-                'ancestors.entity_type.keyword': 'Dataset',
-              },
-            },
-          },
-        },
-      ],
-    },
-  },
-  _source: {
-    includes: ['uuid'],
-  },
-  size: 10000,
-});
-
-const getDescendantsQuery = (uuid: string | string[]): SearchRequest => ({
-  query: {
-    bool: {
-      must_not: {
-        exists: {
-          field: 'descendant_ids',
-        },
-      },
-      must: [
-        {
-          terms: {
-            ancestor_ids: uuid,
-          },
-        },
-        {
-          terms: {
-            'mapped_status.keyword': ['QA', 'Published'],
-          },
-        },
-      ],
-    },
-  },
-  _source: {
-    includes: ['uuid'],
-  },
-  size: 10000,
-});
-
-/**
- * Formats the URL for fetching provenance data
- * @param uuid the UUID of the entity
- * @param entityEndpoint the base URL of the entities API endpoint
- * @returns the URL for fetching provenance data
- */
-function createProvDataURL(uuid: string, entityEndpoint: string) {
-  return `${entityEndpoint}/entities/${uuid}/provenance`;
-}
 /**
  * Fetcher for legacy provenance data
  * This only fetches the ancestors of the entity
@@ -118,22 +22,6 @@ async function getProvData(uuid: string, entityEndpoint: string, groupsToken: st
     throw new Error('Prov API failed to load data');
   }
   return response.json() as Promise<ProvData>;
-}
-
-function nonDestructiveMerge<T extends Record<string, unknown>>(a: T, b: T, keyMod: string): T {
-  const merged = { ...a } as Record<string, unknown>;
-  Object.entries(b).forEach(([key, value]) => {
-    if (merged[key]) {
-      // Confirm that properties are actually different
-      if (JSON.stringify(merged[key]) === JSON.stringify(value)) {
-        return;
-      }
-      merged[`${key}-${keyMod}`] = value;
-    } else {
-      merged[key] = value;
-    }
-  });
-  return merged as T;
 }
 
 /**
@@ -183,6 +71,7 @@ async function getCombinedProvData(
   const [, ...restIds] = descendantIds;
   const combinedProvenance = rest.reduce(
     (acc, curr, idx) => {
+      // Append current descendant's ID to keys if the key itself is a duplicate but its value is not
       const currentId = restIds[idx];
       acc.entity = nonDestructiveMerge(acc.entity, curr.entity, currentId);
       acc.activity = nonDestructiveMerge(acc.activity, curr.activity, currentId);
